requirements.txt
streamlit
phi
openai
google-generativeai
duckduckgo-search
python-dotenv

